+++ app/src/main/res/values/strings.xml
@@ -113,7 +113,7 @@
     <string name="gender">Gender</string>
     <string name="male">Male</string>
     <string name="female">Female</string>
-    <string name="date_of_bith">Date of birth</string>
+    <string name="date_of_birth">Date of birth</string>
     <string name="level">Level</string>
     <string name="your_country">Your country</string>
     <string name="choose_your_location">Choose your location</string>
